requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.1
greenlet==2.0.2
numpy==1.24.2
pandas==1.5.3
paramiko==3.1.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.7
sshtunnel==0.4.0
typing_extensions==4.5.0
